Reduce work chart title size
