+++ app/globals.css
@@ -257,7 +257,7 @@ body {
 }
 
 .work-chart-title {
-  font-size: 0.7rem;
+  font-size: 0.6rem;
   letter-spacing: 0.1em;
 }
